--- src/Assets/Styles/SurveyConstructor.module.scss
@@ -1,5 +1,0 @@
-.wrapper{
-    width: 600px;
-    margin: 0 auto;
-    padding: 25px 0;
-}
