+++ src/Assets/Styles/SurveyForm.module.scss
@@ -1,0 +1,5 @@
+.wrapper{
+    width: 600px;
+    margin: 0 auto;
+    padding: 25px 0;
+}
